+++ src/JlWrap/C.jl
@@ -364,6 +364,7 @@ function init_c()
 end
 
 function __init__()
+    C.CTX.is_initialized || return
     init_c()
 end